requirements.txt
gradio==4.36.1
huggingface_hub==0.20.3
google-generativeai
python-dotenv==1.0.0
matplotlib==3.8.0
networkx==3.1
pillow==10.1.0
reportlab==4.0.7
